requirements.txt
aniso8601
attrs
bcrypt
blinker
certifi
charset-normalizer
click
Flask<3.0.0
Flask-Bcrypt
Flask-Cors
Flask-Mail
flask-restx
Flask-SQLAlchemy
Flask-WTF
greenlet
idna
importlib_resources
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
numpy
pandas
python-dateutil
python-dotenv
pytz
referencing
requests
rpds-py
six
SQLAlchemy
typing_extensions
tzdata
urllib3
Werkzeug<3.0.0
WTForms
tensorflow
keras
Pillow